Move cq_command to impl. Update includes
