+++ tests/tt_metal/tt_metal/dispatch/dispatch_event/test_EnqueueWaitForEvent.cpp
@@ -12,6 +12,8 @@
 #include <tt-metalium/event.hpp>
 #include <tt-metalium/device.hpp>
 
+#include "tt_metal/impl/dispatch/cq_commands.hpp"
+
 namespace tt::tt_metal {
 
 using std::vector;
