+++ tests/tt_metal/tt_metal/dispatch/dispatch_event/test_events.cpp
@@ -10,6 +10,8 @@
 #include <tt-metalium/event.hpp>
 #include <tt-metalium/command_queue.hpp>
 
+#include "tt_metal/impl/dispatch/cq_commands.hpp"
+
 namespace tt::tt_metal {
 
 using std::vector;
