+++ tests/tt_metal/tt_metal/perf_microbenchmark/dispatch/common.h
@@ -10,8 +10,8 @@
 #include <tt-metalium/logger.hpp>
 #include <tt-metalium/host_api.hpp>
 #include <tt-metalium/device.hpp>
-#include <tt-metalium/cq_commands.hpp>
 #include "noc/noc_parameters.h"
+#include "tt_metal/impl/dispatch/cq_commands.hpp"
 
 #include <tt-metalium/hal.hpp>
 #include "llrt.hpp"
