+++ tt_metal/impl/dispatch/debug_tools.cpp
@@ -5,6 +5,7 @@
 #include "debug_tools.hpp"
 
 #include "tt_cluster.hpp"
+#include "tt_metal/impl/dispatch/cq_commands.hpp"
 
 namespace internal {
